pyproject.toml
[tool.poetry]
name = "scrap-thief"
version = "0.1.0"
description = "Scrap Thief"
authors = ["guimassoqueto <gui.massoqueto@gmail.com>"]
readme = "README.md"
packages = [
{include = "*.py"},
{include = "app/**/*.py"},
]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
pydantic = "^2.3.0"
pytz = "^2023.3"
python-dotenv = "^1.0.0"
pika = "^1.3.2"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
